requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic
passlib[bcrypt]
python-jose
alembic
